bigchaindb: basic docker-compose setup
A basic docker-compose setup with mongodb and bigchaindb. Note that we require a properly configured
.bigchaindb file in the project directory. It would be easiest to create some sort of tool that lets us
automatically and easily generate these config files, especially with multi-node setups where we have
to modify the keyring portion each time a new node is added to the network.

Additionally, this current setup only works for single nodes. Multiple nodes would need to run the
command to add their mongodb instance as a replica to the root node. Not sure how I feel about this
since it's supposed to be decentralized and stuff.